backend/requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
pytils==0.3
pytz==2019.3
sqlparse==0.3.0
